ADD public execute methods to AbstractQueryWithFilter(s) with default functionality
REFACTOR applyFiltersToQuery() method to ApplyFilter trait.
REFACTOR AbstractQueryWithFilter's $filter trait to be private.
