+++ tests/Queries/PeopleQueryWithFilters.php
@@ -2,7 +2,6 @@
 
 namespace Sfneal\Datum\Tests\Queries;
 
-use Illuminate\Database\Eloquent\Builder;
 use Sfneal\Datum\Tests\Filters\CityFilter;
 use Sfneal\Datum\Tests\Filters\NameFirstFilterDynamic;
 use Sfneal\Datum\Tests\Filters\NameLastFilter;
@@ -25,22 +24,5 @@ class PeopleQueryWithFilters extends AbstractQueryWithFilters
             'name_last' => NameLastFilter::class,
             'name_first' => NameFirstFilterDynamic::class,
         ];
-    }
-
-    /**
-     * Execute a DB query using filter parameters.
-     *
-     * @return Builder
-     */
-    public function execute(): Builder
-    {
-        // Initialize query
-        $query = $this->builder();
-
-        // Apply filters
-        $query = $this->applyFiltersToQuery($query);
-
-        // Return query
-        return $query;
     }
 }
